+++ web/src/main.ts
@@ -1,6 +1,8 @@
 import { createPinia } from 'pinia'
+import { sendNotification } from '@tauri-apps/api/notification'
 import App from './App.vue'
 import router from './router'
+import { siteMetaData } from '@/constants'
 
 import './assets/main.scss'
 import 'vfonts/OpenSans.css' // Import OpenSans Font
@@ -16,6 +18,14 @@ const app = createApp(App)
 // }
 
 app.use(createPinia()).use(router).mount('#app')
+
+if (import.meta.env.PROD) {
+  sendNotification({
+    title: '🍀二维码生成器',
+    body: '当前仍处于内测阶段，如有问题请反馈管理员。',
+    icon: siteMetaData.Favicon
+  })
+}
 
 faviconLoader()
 naiveUILoader()