+++ web/src/modules/index.ts
@@ -1,1 +1,2 @@
 export { default as Naive } from './naive'
+export * from './downloader'